requirements.txt
pandas==2.1.3
clickhouse-connect==0.6.23
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
schedule==1.2.0
great-expectations==0.18.8